feat(Datagrid): adds FilterPanel stories (#2942)
* feat(web-terminal): adds reduced motion

* feat(web-terminal): adds terminal wrapper component

* chore: merge main

* feat(datagrid): starting to move story files

* feat(datagrid): adds new stories to filter panel

* chore: removed usused and comment
