+++ packages/cloud-cognitive/src/components/Datagrid/Extensions/Filtering/Flyout.stories.js
@@ -1,6 +1,6 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 /**
- * Copyright IBM Corp. 2022, 2022
+ * Copyright IBM Corp. 2022, 2023
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -22,45 +22,13 @@ import { DatagridActions } from '../../utils/DatagridActions';
 import { StatusIcon } from '../../../StatusIcon';
 
 export default {
-  title: `${getStoryTitle(Datagrid.displayName)}/Extensions/Filtering`,
+  title: `${getStoryTitle(Datagrid.displayName)}/Extensions/Filtering/Flyout`,
   component: Datagrid,
   parameters: {
     styles,
     docs: { page: mdx },
   },
 };
-
-const initialFilters = [
-  {
-    id: 'passwordStrength',
-    type: 'checkbox',
-    value: [
-      {
-        id: 'normal',
-        labelText: 'Normal',
-        value: 'normal',
-        selected: true,
-      },
-      {
-        id: 'minor-warning',
-        labelText: 'Minor warning',
-        value: 'minor-warning',
-        selected: false,
-      },
-      {
-        id: 'critical',
-        labelText: 'Critical',
-        value: 'critical',
-        selected: false,
-      },
-    ],
-  },
-  {
-    id: 'role',
-    type: 'radio',
-    value: 'developer',
-  },
-];
 
 const getBatchActions = () => {
   return [
@@ -189,6 +157,7 @@ const FilteringUsage = ({ defaultGridProps }) => {
     {
       columns,
       data,
+      initialState,
       DatagridActions,
       batchActions: true,
       toolbarBatchActions: getBatchActions(),
@@ -198,7 +167,6 @@ const FilteringUsage = ({ defaultGridProps }) => {
       useDenseHeader,
       emptyStateTitle,
       emptyStateDescription,
-      initialState,
     },
     useFiltering
   );
@@ -317,8 +285,8 @@ const filters = [
   },
 ];
 
-export const FilteringUsageStory = prepareStory(FilteringTemplateWrapper, {
-  storyName: 'Filter flyout - batch',
+export const FlyoutBatch = prepareStory(FilteringTemplateWrapper, {
+  storyName: 'Filter flyout with batch update',
   argTypes: {
     gridTitle: ARG_TYPES.gridTitle,
     gridDescription: ARG_TYPES.gridDescription,
@@ -345,67 +313,67 @@ export const FilteringUsageStory = prepareStory(FilteringTemplateWrapper, {
   },
 });
 
-export const FilteringInstantUsageStory = prepareStory(
-  FilteringTemplateWrapper,
-  {
-    storyName: 'Filter flyout - instant',
-    argTypes: {
-      gridTitle: ARG_TYPES.gridTitle,
-      gridDescription: ARG_TYPES.gridDescription,
-      useDenseHeader: ARG_TYPES.useDenseHeader,
-      filterProps: ARG_TYPES.filterProps,
+export const FlyoutInstant = prepareStory(FilteringTemplateWrapper, {
+  storyName: 'Filter flyout with instant update',
+  argTypes: {
+    gridTitle: ARG_TYPES.gridTitle,
+    gridDescription: ARG_TYPES.gridDescription,
+    useDenseHeader: ARG_TYPES.useDenseHeader,
+    filterProps: ARG_TYPES.filterProps,
+  },
+  args: {
+    gridTitle: 'Data table title',
+    gridDescription: 'Additional information if needed',
+    useDenseHeader: false,
+    emptyStateTitle: 'No filters match',
+    emptyStateDescription:
+      'Data was not found with the current filters applied. Change filters or clear filters to see other results.',
+    filterProps: {
+      variation: 'flyout',
+      updateMethod: 'instant',
+      primaryActionLabel: 'Apply',
+      secondaryActionLabel: 'Cancel',
+      flyoutIconDescription: 'Open filters',
+      onFlyoutOpen: action('onFlyoutOpen'),
+      onFlyoutClose: action('onFlyoutClose'),
+      filters,
     },
-    args: {
-      gridTitle: 'Data table title',
-      gridDescription: 'Additional information if needed',
-      useDenseHeader: false,
-      emptyStateTitle: 'No filters match',
-      emptyStateDescription:
-        'Data was not found with the current filters applied. Change filters or clear filters to see other results.',
-      filterProps: {
-        variation: 'flyout',
-        updateMethod: 'instant',
-        primaryActionLabel: 'Apply',
-        secondaryActionLabel: 'Cancel',
-        flyoutIconDescription: 'Open filters',
-        onFlyoutOpen: action('onFlyoutOpen'),
-        onFlyoutClose: action('onFlyoutClose'),
-        filters,
-      },
-    },
-  }
-);
+  },
+});
 
-export const FilteringInitialFiltersUsageStory = prepareStory(
-  FilteringTemplateWrapper,
-  {
-    storyName: 'Filter flyout - with initial filters',
-    argTypes: {
-      gridTitle: ARG_TYPES.gridTitle,
-      gridDescription: ARG_TYPES.gridDescription,
-      useDenseHeader: ARG_TYPES.useDenseHeader,
-      filterProps: ARG_TYPES.filterProps,
+export const FlyoutWithInitialFilters = prepareStory(FilteringTemplateWrapper, {
+  storyName: 'Filter flyout with initial filters',
+  argTypes: {
+    gridTitle: ARG_TYPES.gridTitle,
+    gridDescription: ARG_TYPES.gridDescription,
+    useDenseHeader: ARG_TYPES.useDenseHeader,
+    filterProps: ARG_TYPES.filterProps,
+  },
+  args: {
+    initialState: {
+      filters: [
+        {
+          id: 'role',
+          type: 'radio',
+          value: 'developer',
+        },
+      ],
     },
-    args: {
-      initialState: {
-        filters: initialFilters,
-      },
-      gridTitle: 'Data table title',
-      gridDescription: 'Additional information if needed',
-      useDenseHeader: false,
-      emptyStateTitle: 'No filters match',
-      emptyStateDescription:
-        'Data was not found with the current filters applied. Change filters or clear filters to see other results.',
-      filterProps: {
-        variation: 'flyout',
-        updateMethod: 'instant',
-        primaryActionLabel: 'Apply',
-        secondaryActionLabel: 'Cancel',
-        flyoutIconDescription: 'Open filters',
-        onFlyoutOpen: action('onFlyoutOpen'),
-        onFlyoutClose: action('onFlyoutClose'),
-        filters,
-      },
+    gridTitle: 'Data table title',
+    gridDescription: 'Additional information if needed',
+    useDenseHeader: false,
+    emptyStateTitle: 'No filters match',
+    emptyStateDescription:
+      'Data was not found with the current filters applied. Change filters or clear filters to see other results.',
+    filterProps: {
+      variation: 'flyout',
+      updateMethod: 'instant',
+      primaryActionLabel: 'Apply',
+      secondaryActionLabel: 'Cancel',
+      flyoutIconDescription: 'Open filters',
+      onFlyoutOpen: action('onFlyoutOpen'),
+      onFlyoutClose: action('onFlyoutClose'),
+      filters,
     },
-  }
-);
+  },
+});
